+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.6.12
 - Support Collection Exports
+- Export separate loose meshes does not affect shared meshes (e.g. linked duplicates)
 - Disable UV editor in Level Design workspace (it causes crashes - in the future those may be fixed)
 - Remove Paint Select so Ctrl Left Mouse can use backface-aware shortest path selection